requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.133
distutils
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.33
langchain_openai==0.1.22
langchain_text_splitters==0.2.2
langgraph==0.2.4
numpy==1.24.4
openai==1.41.0
opensearch_py==2.6.0
pandas==2.2.2
playwright==1.45.1
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.1.1
sqlparse==0.5.1
streamlit==1.29.0
typing_extensions==4.12.2
